% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrGA.R
\docType{package}
\name{optimize principal COMponents}
\alias{optimizeCOM}
\title{optimizeCOM: Identification of the optimal number of principal components}
\usage{
optimizeCOM(data, P = 1000, standardize = T, 
       verbose = T)
}
\arguments{
\item{data}{a data frame or matrix. \code{data} has its rows are samples and its columns are genes.}

\item{P}{positive numeric. The number of permutations which is set to 1000 as default.}

\item{standardize}{logical. If your \code{data} are not standardized, just feed \code{T} or \code{TRUE} to this parameter. Default value is \code{T}.}
}
\description{
Determine the optimal number of principal components.
}
\examples{
# Determine the optimal number of components
optimizeCOM(data = exp)
# >> oCEM suggests choosing the optimal number of components is: 9
# >> oCEM also suggests using ICA for your case. 
}
\author{
Quang-Huy Nguyen
}
