% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrGA.R
\docType{package}
\name{overlapping CoExpressed gene Module}
\alias{overlapCEM}
\title{overlapCEM: Detection and correlation analysis of overlapping co-expressed gene modules}
\usage{
overlapCEM(data, clinical, ncomp,
       standardize = T, method, cex.text = 0.7)
}
\arguments{
\item{data}{a data frame or matrix. \code{data} has its rows are samples and its columns are genes.}

\item{clinical}{a data frame or matrix. Input data serve to perform Pearson's correlations between each identified module and each clinical feature. It includes its rows are samples, and its columns are clinical features of your choice.}

\item{ncomp}{positive integer. The optimal number of principal components. It should be >= 2.}

\item{standardize}{logical. If your \code{data} are not standardized, just feed \code{T} or \code{TRUE} to this parameter. Default value is \code{T}.}

\item{method}{Post-processing methods. Allowed values are \code{ICA-FDR}, \code{ICA-Zscore}, or \code{IPCA-FDR}.}

\item{cex.text}{numeric. Change the font size of texts in cells of the heatmap showing correlations between each identified module and each clinical feature. Default value is 0.7.}
}
\description{
Identifies overlapping co-expressed gene modules.
}
\examples{
# Determine the optimal number of components
optimizeCOM(data = exp)
# >> oCEM suggests choosing the optimal number of components is: 9
# >> oCEM also suggests using ICA for your case. 

cem=overlapCEM(data = exp, clinical = clinicalEXP, ncomp = 9,
                        method = 'ICA-Zscore', cex.text = 1.0)


}
\author{
Quang-Huy Nguyen
}
